requirements.txt
six==1.9.0 # for tutum
scrapy
dataset
alembic
psycopg2
xmltodict
sqlalchemy
python-dotenv
